Throw InvalidKeyException if javax.crypto.Mac.init is invoked
